REF: collect DataFrame.__setitem__ tests (#33408)
