+++ pandas/tests/frame/indexing/test_indexing.py
@@ -1921,22 +1921,6 @@ class TestDataFrameIndexing:
         result = df.loc[:, "A"]
         tm.assert_series_equal(result, expected)
 
-    def test_setitem_with_sparse_value(self):
-        # GH8131
-        df = pd.DataFrame({"c_1": ["a", "b", "c"], "n_1": [1.0, 2.0, 3.0]})
-        sp_array = SparseArray([0, 0, 1])
-        df["new_column"] = sp_array
-        tm.assert_series_equal(
-            df["new_column"], pd.Series(sp_array, name="new_column"), check_names=False
-        )
-
-    def test_setitem_with_unaligned_sparse_value(self):
-        df = pd.DataFrame({"c_1": ["a", "b", "c"], "n_1": [1.0, 2.0, 3.0]})
-        sp_series = pd.Series(SparseArray([0, 0, 1]), index=[2, 1, 0])
-        df["new_column"] = sp_series
-        exp = pd.Series(SparseArray([1, 0, 0]), name="new_column")
-        tm.assert_series_equal(df["new_column"], exp)
-
     def test_setitem_with_unaligned_tz_aware_datetime_column(self):
         # GH 12981
         # Assignment of unaligned offset-aware datetime series.
